gru-prediction-service/requirements.txt
pandas
numpy
torch
scikit-learn
joblib
google-cloud-bigquery
google-cloud-storage
tzdata
db-dtypes